requirements.txt
Django==4.1.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
